use of null transformer in datetime and numerical
