+++ rdt/transformers/__init__.py
@@ -2,7 +2,8 @@ from rdt.transformers.base import BaseTransformer
 from rdt.transformers.category import CategoricalTransformer
 from rdt.transformers.datetime import DateTimeTransformer
 from rdt.transformers.null import NullTransformer
-from rdt.transformers.number import NumericalTransformer
+from rdt.transformers.numerical import NumericalTransformer
+from rdt.transformers.positive_number import PositiveNumericalTransformer
 
 __all__ = [
     'BaseTransformer',
@@ -10,4 +11,5 @@ __all__ = [
     'DateTimeTransformer',
     'NumericalTransformer',
     'NullTransformer',
+    'PositiveNumericalTransformer',
 ]
